project_mercury/rtm/src/main/java/com/smartsparrow/rtm/subscription/courseware/assetremoved/AssetRemovedRTMEventDecoratorImpl.java
package com.smartsparrow.rtm.subscription.courseware.assetremoved;

import com.smartsparrow.courseware.data.CoursewareElementType;
import com.smartsparrow.pubsub.data.RTMEvent;

public class AssetRemovedRTMEventDecoratorImpl extends AssetRemovedRTMEventDecorator {

    public AssetRemovedRTMEventDecoratorImpl(RTMEvent rtmEvent) {
        super(rtmEvent);
    }

    public String getName(CoursewareElementType elementType) {
        return elementType.toString() + "_" + super.getName();
    }

    public Boolean equalsTo(RTMEvent rtmEvent) {
        return super.equalsTo(rtmEvent);
    }

    public String getLegacyName() {
        return super.getLegacyName();
    }
}
